+++ Editor/DLLs/ContentToolsAPI.cs
@@ -11,10 +11,23 @@ namespace Editor.DLLs
     {
         private const string _contentToolsDll = "ContentTools.dll";
 
-        [DllImport(_contentToolsDll)]
+        [DllImport(_contentToolsDll)] // Modify entry point
         private static extern void CreatePrimitiveMesh([In, Out] SceneData data, PrimitiveInitInfo info);
 
-        public static void CreatePrimitiveMesh(Geometry geometry, PrimitiveInitInfo info)
+        [DllImport(_contentToolsDll)] // Modify entry point
+        private static extern void ImportFbx(string file, [In, Out] SceneData data);
+
+        public static void CreatePrimitiveMesh(Geometry geometry, PrimitiveInitInfo info) =>
+            GeometryFromSceneData(
+                geometry,
+                (sceneData) => CreatePrimitiveMesh(sceneData, info),
+                $"Failed to create {info.Type} primitive mesh."
+            );
+
+        public static void ImportFbx(string file, Geometry geometry) =>
+            GeometryFromSceneData(geometry, (sceneData) => ImportFbx(file, sceneData), $"Failed to import from FBX file: {file}");
+
+        private static void GeometryFromSceneData(Geometry geometry, Action<SceneData> sceneDataGenerator, string failureMessage)
         {
             Debug.Assert(geometry is not null);
 
@@ -23,7 +36,7 @@ namespace Editor.DLLs
             try
             {
                 sceneData.ImportSettings.FromContentSettings(geometry);
-                CreatePrimitiveMesh(sceneData, info);
+                sceneDataGenerator(sceneData);
 
                 Debug.Assert(sceneData.Data != IntPtr.Zero && sceneData.DataSize > 0);
 
@@ -33,7 +46,7 @@ namespace Editor.DLLs
             }
             catch (Exception ex)
             {
-                Logger.LogAsync(LogLevel.ERROR, $"Failed to create {info.Type} primitive mesh.");
+                Logger.LogAsync(LogLevel.ERROR, failureMessage);
                 Debug.WriteLine(ex.Message);
             }
         }