+++ Editor/GameCode/VisualStudio.cs
@@ -239,9 +239,10 @@ namespace Editor.GameCode
 
             CallOnSTAThread(() => {
 
+                _vsInstance.MainWindow.Visible = showWindow;
+                
                 if (!_vsInstance.Solution.IsOpen) _vsInstance.Solution.Open(project.Solution);
 
-                _vsInstance.MainWindow.Visible = showWindow;
                 _vsInstance.Events.BuildEvents.OnBuildProjConfigBegin += OnBuildSolutionBegin;
                 _vsInstance.Events.BuildEvents.OnBuildProjConfigDone += OnBuildSolutionDone;
             });