+++ Editor/Utilities/ContentHelper.cs
@@ -1,4 +1,6 @@
-﻿using System.Diagnostics;
+﻿using Editor.Content;
+using Editor.Content.ContentBrowser;
+using System.Diagnostics;
 using System.IO;
 using System.Security.Cryptography;
 using System.Text;
@@ -41,6 +43,81 @@ namespace Editor.Utilities
             }
 
             return false;
+        }
+
+        public static async Task ImportFilesAsync(string[] files, string destination)
+        {
+            try
+            {
+                Debug.Assert(!string.IsNullOrEmpty(destination));
+
+                ContentWatcher.EnableFileWatcher(false);
+
+                var tasks = files.Select(async file => await Task.Run(() =>
+                {
+                    Import(file, destination);
+                }));
+
+                await Task.WhenAll(tasks);
+            }
+            catch(Exception ex)
+            {
+                Debug.WriteLine($"Failed to import files to {destination}");
+                Debug.WriteLine(ex.Message);
+            }
+            finally
+            {
+                ContentWatcher.EnableFileWatcher(true);
+            }
+        }
+
+        private static void Import(string file, string destination)
+        {
+            Debug.Assert(!string.IsNullOrEmpty(file));
+
+            if (IsDirectory(file)) return;
+            if(!destination.EndsWith(Path.DirectorySeparatorChar)) destination += Path.DirectorySeparatorChar;
+
+            var name = Path.GetFileNameWithoutExtension(file).ToLower();
+            var ext = Path.GetExtension(file).ToLower();
+
+            Asset asset = null;
+
+            switch(ext)
+            {
+                case ".fbx":
+                    asset = new Geometry();
+                    break;
+                case ".bmp": break;
+                case ".png": break;
+                case ".jpg": break;
+                case ".jpeg": break;
+                case ".tiff": break;
+                case ".tif": break;
+                case ".tga": break;
+                case ".waw": break;
+                case ".ogg": break;
+                default:
+                    break;
+            }
+
+            if(asset is not null)
+            {
+                Import(asset, name, file, destination);
+            }
+        }
+
+        private static void Import(Asset asset, string name, string file, string destination)
+        {
+            Debug.Assert(asset is not null);
+
+            asset.FullPath = destination + name + Asset.AssetFileExtension;
+
+            if (!string.IsNullOrEmpty(file)) asset.Import(file);
+
+            asset.Save(asset.FullPath);
+
+            return;
         }
     }
 }