+++ Changelog.md
@@ -7,6 +7,8 @@
  - Added visualization of importance sampling for specular
 
 ## Editor
+ - Drag-and-drop importer
+ - Import FBX files
  - Content browser
  - Asset registry
  - Dynamic switch between projects